feat: update message handling
